Moved opm/core/utility to opm/grid/utility.
