+++ opm/grid/utility/StopWatch.cpp
@@ -36,7 +36,7 @@
 #include "config.h"
 #endif
 #include <boost/date_time/posix_time/posix_time.hpp>
-#include <opm/core/utility/StopWatch.hpp>
+#include <opm/grid/utility/StopWatch.hpp>
 #include <opm/grid/utility/ErrorMacros.hpp>
 
 namespace Opm
