+++ opm/grid/utility/WachspressCoord.hpp
@@ -20,7 +20,7 @@
 #ifndef OPM_WACHSPRESSCOORD_HEADER_INCLUDED
 #define OPM_WACHSPRESSCOORD_HEADER_INCLUDED
 
-#include <opm/core/utility/SparseTable.hpp>
+#include <opm/grid/utility/SparseTable.hpp>
 #include <vector>
 
 struct UnstructuredGrid;
